Steve vs Rancher types, types.js is only for type names
