+++ config/schema.js
@@ -1,0 +1,18 @@
+export const COLLECTION_TYPES = {
+  array: 'array',
+  map:   'map',
+};
+
+export const PRIMITIVE_TYPES = {
+  string:    'string',
+  multiline: 'multiline',
+  masked:    'masked',
+  password:  'password',
+  float:     'float',
+  int:       'int',
+  date:      'date',
+  blob:      'blob',
+  boolean:   'boolean',
+  json:      'json',
+  version:   'version',
+};
