+++ models/service.js
@@ -1,0 +1,24 @@
+export const DEFAULT_SERVICE_TYPES = [
+  {
+    id:               'ClusterIP',
+    translationLabel: 'serviceTypes.clusterip'
+  },
+  {
+    id:               'ExternalName',
+    translationLabel: 'serviceTypes.externalname'
+  },
+  {
+    id:               'Headless',
+    translationLabel: 'serviceTypes.headless'
+  },
+  {
+    id:               'LoadBalancer',
+    translationLabel: 'serviceTypes.loadbalancer'
+  },
+  {
+    id:               'NodePort',
+    translationLabel: 'serviceTypes.nodeport'
+  },
+];
+
+export default {};
